add user id to campaign migration
